Move `Readable` implementation for `FromStr`
There is almost no advantage to separate definition and implementation,
so move the implementation to the definition site and remove `read`
module completely. `read` module doesn't have any exported items so
removing `read` shouldn't cause any compatiblility problems --- except
that someone meaninglessly import `read` module itself. (Of course that
case hardly occurs, and even if it occurs, simply removing the import
will fix the issue.)
